requirements.txt
requests~=2.26.0
Flask~=2.0.1
Pillow~=8.3.1
pandas~=1.3.2
docx~=0.2.4